requirements/base.txt
Django==3.2.9
python-decouple==3.5
django-environ==0.8.1
django-model-utils==4.2.0
django-import-export==2.6.1
django-cors-headers==3.10.0
djangorestframework==3.12.4
dj-rest-auth==2.1.11
djangorestframework-simplejwt==5.0.0
djangorestframework-filters==1.0.0.dev0
django-allauth==0.46.0
PyJWT==2.3.0
